Update patch set 1
Patch Set 1:

Finally took some time to look at this. Assuming the data collected for core1 at low freq is just an anomaly, and that the power cost to add a core is the same as the power cost to add the first core you values more or less seem logical. 

My question i guess is: Is this what that portion is of the overlay is supposed to contain? The power draw of adding 1 core at the given freq? Or should ti be the full power draw of all cores at that freq?

I'd be curious to see similar testing on a few other devices that we more or less trust the current power_profile for as a comparison.

Patch-set: 1
Label: Code-Review=0