Update patch set 1
Patch Set 1:

Good suggestion to look at another device. I tried this on victara and unfortunately, the battery reports in the input current from the charger (aka, it's constant no matter what load I put on the system).

I wonder if we're lucky in how Z00A's hardware is laid-out that the CPU draws power from the input before it reaches the battery in some way?

Patch-set: 1